Fix shellcheck warnings and add check to pipeline

+++ scripts/ytccf.sh
@@ -24,8 +24,8 @@ set -o pipefail
 set -o errexit
 set -o nounset
 
-FILTERS=''
-MAKE_TABLE='ytcc --output table --truncate $((2 * TERM_COLS / 3 - 3)) list --attributes id,title,publish_date,duration,playlists'
+declare -r -x THUMBNAIL_DIR=$HOME/.cache/ytccf/thumbnails
+FILTERS=()
 KEY_BINDINGS="
         tab: select/deselect
       enter: play video(s)
@@ -36,14 +36,21 @@ KEY_BINDINGS="
       alt-u: mark last watched video as unwatched
       alt-h: show help"
 
-THUMBNAILS=1
-declare -r -x THUMBNAIL_DIR=$HOME/.cache/ytccf/thumbnails
+# shellcheck disable=SC2016
+MAKE_TABLE='ytcc --output table --truncate $((2 * $(tput cols) / 3 - 3)) list --attributes id,title,publish_date,duration,playlists'
+
+if command -v ueberzug &> /dev/null || [[ $TERM == "xterm-kitty" ]]; then
+    THUMBNAILS=1
+else
+    THUMBNAILS=0
+fi
 
 function draw_preview() { :; }
-function clear_preview() { :; } 
-function init_preview() { :; } 
-function finalize_preview() { :; } 
+function clear_preview() { :; }
+function init_preview() { :; }
+function finalize_preview() { :; }
 function fetch_thumbnails() { :; }
+function calculate_preview_size() { :; }
 
 function check_cmd() {
     if ! command -v "$1" &> /dev/null; then
@@ -91,31 +98,31 @@ while [[ $# -gt 0 ]]; do
     key="$1"
     case $key in
     -p | --playlists)
-        FILTERS="$FILTERS -p '$2'"
+        FILTERS+=(-p "$2")
         shift
         shift
         ;;
     -c | --tags)
-        FILTERS="$FILTERS -c '$2'"
+        FILTERS+=( -c "$2")
         shift
         shift
         ;;
     -s | --since)
-        FILTERS="$FILTERS -s '$2'"
+        FILTERS+=( -s "$2")
         shift
         shift
         ;;
     -t | --till)
-        FILTERS="$FILTERS -t '$2'"
+        FILTERS=(-t "$2")
         shift
         shift
         ;;
     -w | --watched)
-        FILTERS="$FILTERS -w"
+        FILTERS=(-w)
         shift
         ;;
     -u | --unwatched)
-        FILTERS="$FILTERS -u"
+        FILTERS=(-u)
         shift
         ;;
     -h | --help)
@@ -127,7 +134,7 @@ while [[ $# -gt 0 ]]; do
         shift
         ;;
     --clear-thumbnails)
-        rm -r "$THUMBNAIL_DIR" && echo "Successfully cleared thumbnail cache"
+        [[ -d "$THUMBNAIL_DIR" ]] && rm -r "$THUMBNAIL_DIR" && echo "Successfully cleared thumbnail cache"
         exit
         ;;
     *)
@@ -137,97 +144,92 @@ while [[ $# -gt 0 ]]; do
         ;;
     esac
 done
-MAKE_TABLE="$MAKE_TABLE $FILTERS"
+MAKE_TABLE="$MAKE_TABLE ${FILTERS[*]}"
 
 check_cmd ytcc
 check_cmd fzf
-check_cmd stty
 
 if [[ $THUMBNAILS -eq 1 ]]; then
     check_cmd curl
+    check_cmd stty
 
     function fetch_thumbnails() {
-        curl_args=""
-        for line in $(ytcc --output xsv list --attributes id,thumbnail_url $FILTERS); do
-            arr=(${line/,/ })
-            ((${#arr[@]} > 1)) \
-                && ! [[ -e $THUMBNAIL_DIR/${arr[0]} ]] \
-                && curl_args+=" -o $THUMBNAIL_DIR/${arr[0]} ${arr[1]}"
+        local -a curl_args=()
+        for line in $(ytcc --output xsv list --attributes id,thumbnail_url "${FILTERS[@]}"); do
+            read -r -a arr <<< "${line/,/ }"
+            if ((${#arr[@]} > 1)) && ! [[ -e $THUMBNAIL_DIR/${arr[0]} ]]; then
+                curl_args+=( -o "$THUMBNAIL_DIR/${arr[0]}" "${arr[1]}")
+            fi
         done
         mkdir -p "$THUMBNAIL_DIR"
-        if [[ $curl_args != "" ]]; then
+        if (( ${#curl_args[@]} > 0 )); then
             echo INFO: Fetching thumbnails
-            curl -L --silent $curl_args
+            curl -L --silent "${curl_args[@]}"
         fi
+    }
+
+    function calculate_preview_size() {
+        local term_lines term_cols
+        read -r term_lines term_cols < <(stty size </dev/tty)
+        X=$((2 * term_cols / 3))
+        Y=3
+        LINES=$((term_lines / 2 - 3))
+        COLUMNS=$((term_cols / 3 ))
     }
 
     if [[ $TERM == "xterm-kitty" ]]; then
         function draw_preview {
-            read TERM_LINES TERM_COLS < <(</dev/tty stty size)
-            X=$((2 * TERM_COLS / 3))
-            Y=3
-            LINES=$((TERM_LINES / 2 - 3))
-            COLUMNS=$((TERM_COLS / 3 ))
-            export TERM_LINES TERM_COLS
-
+            calculate_preview_size
             kitty icat --transfer-mode file -z=-1 --place=${COLUMNS}x${LINES}@${X}x${Y} --scale-up "${@}"
         }
     else
         check_cmd ueberzug
-        declare -r -x UEBERZUG_FIFO="$(mktemp --dry-run --suffix "fzf-$$-ueberzug")"
+        UEBERZUG_FIFO="$(mktemp --dry-run --suffix "fzf-$$-ueberzug")"
+        declare -r -x UEBERZUG_FIFO
         declare -r -x PREVIEW_ID="preview"
 
         function draw_preview {
-            read TERM_LINES TERM_COLS < <(</dev/tty stty size)
-            X=$((2 * TERM_COLS / 3))
-            Y=3
-            LINES=$((TERM_LINES / 2 - 3))
-            COLUMNS=$((TERM_COLS / 3 ))
-            export TERM_LINES TERM_COLS
-
-            >"${UEBERZUG_FIFO}" declare -A -p cmd=( \
-                [action]=add [identifier]="${PREVIEW_ID}" \
+            calculate_preview_size
+            local -A cmd=( \
+                [action]=add
+                [identifier]="${PREVIEW_ID}" \
                 [x]="${X}" [y]="${Y}" \
                 [width]="${COLUMNS}" [height]="${LINES}" \
                 [scaler]=fit_contain [scaling_position_x]=0.5 [scaling_position_y]=0.5 \
-                [path]="${@}")
-                # add [synchronously_draw]=True if you want to see each change
+                [path]="${@}"
+            )
+
+            declare -p cmd >"${UEBERZUG_FIFO}"
         }
 
         function clear_preview {
-            >"${UEBERZUG_FIFO}" declare -A -p cmd=( \
+            local -A cmd=( \
                 [action]=remove [identifier]="${PREVIEW_ID}" \
             )
+            declare -p cmd >"${UEBERZUG_FIFO}"
         }
 
         function init_preview {
             mkfifo "${UEBERZUG_FIFO}"
-            <"${UEBERZUG_FIFO}" \
-                ueberzug layer --parser bash --silent &
+            ueberzug layer --parser bash --silent <"${UEBERZUG_FIFO}" &
+
             # prevent EOF
-            3>"${UEBERZUG_FIFO}" \
-                exec
+            exec 3>"${UEBERZUG_FIFO}"
         }
 
         function finalize_preview {
-            3>&- \
-                exec
-            &>/dev/null \
-                rm "${UEBERZUG_FIFO}"
-            &>/dev/null \
-                kill "$(jobs -p)"
+            exec 3>&-
+            rm "${UEBERZUG_FIFO}" &>/dev/null
+            kill "$(jobs -p)" &>/dev/null
         }
     fi
 fi
-
 
 init_preview
 trap finalize_preview EXIT
 fetch_thumbnails
 
-read TERM_LINES TERM_COLS < <(</dev/tty stty size)
-export TERM_COLS TERM_LINES
-export -f draw_preview clear_preview fetch_thumbnails
+export -f draw_preview clear_preview fetch_thumbnails calculate_preview_size
 
 eval "$MAKE_TABLE" |
     SHELL=/usr/bin/bash fzf --preview "draw_preview $THUMBNAIL_DIR/{1}; ytcc --output xsv --separator ';' list --watched --unwatched -a description -i {1}" \
